[Feat] 다녀온 전시 별점 기능 구현

+++ app/src/main/java/com/mobile/seoulmoa_zip/data/ExhibitionEntity.kt
@@ -19,5 +19,5 @@ data class ExhibitionEntity(
     val link: String?, // 바로가기링크
     var isLiked: Boolean?,
     var isVisited: Boolean?,
-    var score : Int?
+    var score: Float?
 )
